+++ package.json
@@ -8,6 +8,7 @@
     "@mui/icons-material": "^5.2.1",
     "@mui/material": "^5.2.3",
     "@mui/styles": "^5.2.3",
+    "@netlify/functions": "^1.6.0",
     "firebase": "^7.21.1",
     "lodash": "^4.17.20",
     "node-fetch": "2.6.11",
@@ -24,13 +25,13 @@
     "@trivago/prettier-plugin-sort-imports": "^1.3.0",
     "@types/jest": "^24.0.0",
     "@types/lodash": "^4.14.161",
-    "@types/node": "^12.0.0",
+    "@types/node": "^14.18.53",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-router-dom": "^5.1.5",
     "cra-append-sw": "^2.7.0",
     "husky": "^4.3.0",
-    "typescript": "^4.7.4"
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "start": "react-scripts start",
